Make changes for perl 5.22 and before
Perl 5.22 and earlier versions required that if an attribute has a writer, then it can only have a read-only accessor. So things started failing on Travis-CI. So the solution adopted is to create a method named filehandle and instead create a reader method called _get_filehandle. This resolves the issue I think.